+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## [Unreleased]
 ### Changed
 - Changed `decryptAES` to use const ref instead of changing original data
-- Start to move to Crypto++
+- Base64 functions using Crypto++
+- AES encryption using Crypto++
 
 ## [1.1.4] - 01-02-2017
 ### Changed